requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.6.0
certifi==2022.12.7
click==8.1.3
Django==4.1.1
django-ckeditor==6.4.2
django-debug-toolbar==3.6.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.6
sqlparse==0.4.2
tomli==2.0.1
